install react native paper & move ProfileCard to component
