+++ mobile-app/package.json
@@ -18,6 +18,7 @@
     "expo": "~45.0.0",
     "expo-barcode-scanner": "~11.3.0",
     "expo-status-bar": "~1.3.0",
+    "expo-updates": "~0.13.4",
     "react": "17.0.2",
     "react-dom": "17.0.2",
     "react-native": "0.68.2",
@@ -25,8 +26,7 @@
     "react-native-qrcode-svg": "^6.1.2",
     "react-native-svg": "12.3.0",
     "react-native-web": "0.17.7",
-    "rn-qr-generator": "^1.2.0",
-    "expo-updates": "~0.13.4"
+    "rn-qr-generator": "^1.2.0"
   },
   "devDependencies": {
     "@babel/core": "^7.12.9"
